pyproject.toml
[project]
name = "generative-ai-assistance-for-literature-reviews-in-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.1.3",
    "langchain-chroma>=0.2.2",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.12",
    "openai>=1.70.0",
    "pypdf>=5.4.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]
